refactor: reduce ambiguity between sorter/matcher
